requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
aiosqlite>=0.19.0
httpx>=0.25.0
python-dotenv>=1.0.0
click>=8.1.0
anthropic>=0.39.0
openai>=1.50.0
sse-starlette>=1.8.0
